chore: cleanups pre brian, NodeId everywhere

+++ crates/poe_tree/src/type_wrappings.rs
@@ -1,4 +1,8 @@
 //!$ crates/poe_tree/src/type_wrappings.rs
-pub type GroupId = u32;
-pub type NodeId = u32;
-pub type EdgeId = u32;
+// pub type GroupId = u32;
+// pub type NodeId = u32;
+// pub type EdgeId = u32;
+
+pub type GroupId = u16;
+pub type NodeId = u16;
+pub type EdgeId = u16;
